fix: add separate /:orderId pages and created unique routes
