+++ src/components/order-details/styles.module.css
@@ -1,0 +1,82 @@
+.main {
+  max-width: 1240px;
+  max-height: calc(100vh - 88px);
+  width: 100%;
+  margin: 0 auto;
+  background-color: #131316;
+}
+.section {
+  margin: 0 auto;
+  width: 640px;
+  min-height: 600px;
+  text-align: left;
+}
+.notFound {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+}
+.number {
+  text-align: center;
+}
+.status {
+  color: var(--colors-interface-success);
+}
+.ingredients {
+  box-sizing: border-box;
+  width: 100%;
+  list-style-type: none;
+  padding: 0 30px 0 0;
+  margin-bottom: 40px;
+  display: flex;
+  gap: 16px;
+  flex-direction: column;
+  max-height: 310px;
+  height: fit-content;
+  overflow-y: scroll;
+  scroll-snap-type: both mandatory;
+}
+.ingredients::-webkit-scrollbar {
+  width: 12px;
+  z-index: 2;
+}
+.ingredients::-webkit-scrollbar-track {
+  background:var(--background-element);
+  border-radius: 6px;
+}
+.ingredients::-webkit-scrollbar-thumb {
+  background: #8585AD;
+  border: 3px solid var(--background-element);
+  border-radius: 6px;
+}
+.listItem {
+  scroll-snap-align: start;
+  height: 64px;
+}
+.card {
+  height: 64px;
+  width: 100%;
+  display: grid;
+  grid-template-columns: 64px 10fr 4fr;
+  gap: 16px;
+}
+.gridCenteredItem {
+  align-self: center;
+}
+.cost {
+  align-self: center;
+  display: flex;
+  justify-content: flex-end;
+  gap: 4px;
+  align-items: center;
+}
+.thumbnail {
+  height: 64px;
+  width: 64px;
+}
+.timeAndTotalWrap {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
