+++ src/components/profile-menu/styles.module.css
@@ -1,0 +1,13 @@
+.navigation {
+  display: flex;
+  flex-direction: column;
+}
+.link {
+  height: 64px;
+  text-decoration: none;
+  line-height: 64px;
+  color: var(--text-inactive-color)
+}
+.activeLink {
+  color: var(--text-primary-color)
+}
