+++ src/pages/profile-edit-page.module.css
@@ -9,19 +9,7 @@
   padding: 120px 0 20px 0;
   margin: 0 auto;
 }
-.navigation {
-  display: flex;
-  flex-direction: column;
-}
-.link {
-  height: 64px;
-  text-decoration: none;
-  line-height: 64px;
-  color: var(--text-inactive-color)
-}
-.activeLink {
-  color: var(--text-primary-color)
-}
+
 .orders {
   margin-top: -36px;
   max-height: calc(100vh - 320px);
