+++ src/pages/profile-order-details.jsx
@@ -1,0 +1,8 @@
+import styles from './profile-order-details.module.css';
+import { OrderDetails } from '../components/order-details'
+
+export const ProfileOrderDetails = () => (
+  <main className={styles.main}>
+    <OrderDetails />
+  </main>
+)
