+++ src/pages/profile-order-details.module.css
@@ -1,4 +1,5 @@
 .main {
+  padding-top: 120px;
   max-width: 1240px;
   max-height: calc(100vh - 88px);
   width: 100%;
