+++ src/utils/helpers.js
@@ -45,17 +45,29 @@ const daysAgoToString = days => days < 0
   ? `${days} дня(-ей) назад`
   :'Ошибка в вычислении времени'
 
-export const orderDateAgoToString = (date1, date2 = new Date()) => {
-  date2.setHours(0);
-  date2.setMinutes(0);
-  date2.setSeconds(0);
+export const orderDateAgoToString = date => {
+  const today = new Date();
+  today.setHours(0);
+  today.setMinutes(0);
+  today.setSeconds(0);
 
-  const daysAgo = Math.ceil((date2 - date1) / (60 * 60 * 24 * 1000));
+  const daysAgo = Math.ceil((today - date) / (60 * 60 * 24 * 1000));
 
   const days = daysAgoToString(daysAgo);
-  const hours = date1.getHours();
-  const minutes = date1.getMinutes();
-  const timeShift = `i-GMT${date1.getTimezoneOffset() / 60}`
+  const hours = date.getHours();
+  const minutes = date.getMinutes();
+  const timeShift = `i-GMT${date.getTimezoneOffset() / 60}`
 
   return `${days}, ${hours}:${minutes} ${timeShift}`
 }
+
+export const countIngredients = ingredients => ingredients
+  .reduce((acc, item) => {
+    const currentItemIndexInAcc = acc.findIndex((i) => i._id === item._id);
+
+    if (currentItemIndexInAcc < 0) return [...acc, { count: 1, ...item }]
+    
+    acc[currentItemIndexInAcc]['count'] += 1;
+
+    return acc;
+  }, [])
